Added:In Customtextfield close icon for text clear is added & logout is working.
