+++ lib/cart/bloc/add_to_cart_state.dart
@@ -22,4 +22,3 @@ class CartError extends AddToCartState {
 
   CartError(this.message);
 }
-
